Add JRuby compatibility
Since version 1.0.0 logasm became unusable in JRuby because
of Oj dependency (it's a C-extension and not compatible with JRuby).

This commit adds a compatibility with JRuby by using Java-specific
JSON engine JrJackson which claims to be faster than Oj on long run.

Also, for JRuby we don't have to manually convert time object into
desired format, because JrJackson supports custom time formats.