Automatically publish package on version bump
Previously the package needed to be manually published by somebody with
permissions. This is a hasstle if the changes were authored by somebody without
permissions. A Github Action will automatically publish a new version and add a
tag when changes are merged to master that increase the version number in
gemspec. If the version in unchanged when changes are merged then the action
fails but does so silently.

CHAN-1774